+++ example/test_driver/app.dart
@@ -1,5 +1,4 @@
-import 'package:example/main.dart';
-import 'package:flutter/widgets.dart';
+import 'package:example/main.dart' as app;
 import 'package:flutter_driver/driver_extension.dart';
 
 void main() {
@@ -8,5 +7,5 @@ void main() {
 
   // Call the `main()` function of your app or call `runApp` with any widget you
   // are interested in testing.
-  runApp(MyApp());
+  app.main();
 }